chore: convert maze import script to JS
