+++ package.json
@@ -3,14 +3,14 @@
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
-    "prestart": "node ./scripts/update-maze-import.ts",
+    "prestart": "node ./scripts/update-maze-import.js",
     "start": "expo start",
     "reset-project": "node ./scripts/reset-project.js",
-    "preandroid": "node ./scripts/update-maze-import.ts",
+    "preandroid": "node ./scripts/update-maze-import.js",
     "android": "expo start --android",
-    "preios": "node ./scripts/update-maze-import.ts",
+    "preios": "node ./scripts/update-maze-import.js",
     "ios": "expo start --ios",
-    "preweb": "node ./scripts/update-maze-import.ts",
+    "preweb": "node ./scripts/update-maze-import.js",
     "web": "expo start --web",
     "lint": "expo lint",
     "test": "jest"
